+++ web/src/components/entry-dialog/EntryForm.tsx
@@ -1,3 +1,4 @@
+import { useMutation, useQuery } from "@apollo/client";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import { LoadingButton } from "@mui/lab";
 import {
@@ -27,6 +28,9 @@ import {
   GetMySettingsDocument,
   UpdateSettingsDocument,
 } from "../../graphql/generated/graphql";
+import { JiraIntegrationAlert } from "../../jira/components/JiraIntegrationAlert";
+import JiraIssueComboBox from "../../jira/components/JiraIssueComboBox";
+import { useIsJiraAuthenticated } from "../../jira/jiraApi";
 import usePreferSetRemainingHours from "../user-preferences/usePreferSetRemainingHours";
 import BigDeleteEntryButton from "./BigDeleteEntryButton";
 import DimensionComboBox from "./DimensionComboBox";
@@ -34,10 +38,6 @@ import DurationSlider from "./DurationSlider";
 import ResponsiveDatePicker from "./ResponsiveDatePicker";
 import WorkdayHours from "./WorkdayHours";
 import useEntryForm, { EntryFormSchema } from "./useEntryForm";
-import { useIsJiraAuthenticated } from "../../jira/jiraApi";
-import JiraIssueComboBox from "../../jira/components/JiraIssueComboBox";
-import { JiraIntegrationAlert } from "../../jira/components/JiraIntegrationAlert";
-import { useMutation, useQuery } from "@apollo/client";
 
 export type EntryFormProps = {
   form: UseFormReturn<EntryFormSchema>;
